pyproject.toml
[project]
name = "fluxium-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "granian>=2.7.2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-jaeger-thrift>=1.11.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-aiohttp-client>=0.60b1",
    "opentelemetry-instrumentation-asgi>=0.60b1",
    "opentelemetry-instrumentation-pymongo>=0.60b1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "orjson>=3.11.7",
    "prometheus-client>=0.24.1",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "redis>=7.2.1",
]

